+++ src/user.c
@@ -1029,10 +1029,3 @@ struct sockaddr *raw_client_ip(Client *client)
 	}
 	return NULL;
 }
-
-/* Get the max-channels-per-user setting for this client */
-int get_max_channels_per_user(Client *client)
-{
-	FloodSettings *settings = get_floodsettings_for_user(client, FLD_MAXCHANNELSPERUSER);
-	return (settings && settings->limit[FLD_MAXCHANNELSPERUSER]) ? settings->limit[FLD_MAXCHANNELSPERUSER] : iConf.maxchannelsperuser;
-}